fixed issue#1

+++ src/game.h
@@ -4,12 +4,12 @@
 typedef struct {
     unsigned int prevTime;
     unsigned int currentTime;
-    float deltaTime;
-    float deltaTick;
-    float physicsTime;
-    float physicsIter;
-    float renderTime;
-    float renderIter;
+    double deltaTime;
+    double deltaTick;
+    double physicsTime;
+    double physicsIter;
+    double renderTime;
+    double renderIter;
 } Timer;
 
 typedef struct {
@@ -59,7 +59,7 @@ typedef struct {
 
 Game *get_game();
 Background *get_background();
-float get_tick_delta();
+double get_tick_delta();
 void init_game();
 GameState get_game_state();
 void set_game_state(GameState state);
